Use .to fp32 insead of .float

+++ utils/models/raiflow_utils.py
@@ -154,9 +154,9 @@ def run_raiflow_model_training(
                     encoder_hidden_states=prompt_embeds,
                     timestep=sigmas,
                     return_dict=False,
-                )[0].float()
+                )[0].to(dtype=torch.float32)
 
-            noisy_model_input = noisy_model_input.float()
+            noisy_model_input = noisy_model_input.to(dtype=torch.float32)
             noisy_model_input, target, self_correct_count = get_self_corrected_targets(
                 noisy_model_input=noisy_model_input,
                 target=target,
@@ -199,8 +199,8 @@ def run_raiflow_model_training(
     else:
         raise RuntimeError(f'Prediction type {config["prediction_type"]} is not implemented for {config["model_type"]}')
 
-    model_pred = model_pred.float()
-    target = target.float().detach()
+    model_pred = model_pred.to(dtype=torch.float32)
+    target = target.to(dtype=torch.float32).detach()
 
     if config["loss_weighting"] == "sigma_sqrt":
         sigma_sqrt = sigmas.sqrt().clamp(min=0.1, max=None)
